requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
click==8.1.8
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pydantic==2.10.4
pydantic-settings==2.1.0
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-jose==3.3.0
cryptography==41.0.7
passlib==1.7.4
bcrypt==4.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
